% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase.R
\name{camelCase}
\alias{camelCase}
\title{camelCase is cool}
\usage{
camelCase()
}
\value{
single-element \code{character} vector with an educated opinon on
camelCase
}
\description{
This function doesn't like the \code{camel_case_linter} because it's not
called \code{camelCaseLinter}
}
\examples{
camelCase()

}
